Ajustes no H2 e na injeçao de dependencia

+++ src/main/resources/db/migration/V1_001__create_carro.sql
@@ -1,5 +1,5 @@
 CREATE TABLE if not exists CARRO (
-  id integer NOT NULL AUTO_INCREMENT,
+  id bigint NOT NULL AUTO_INCREMENT,
   nome varchar(200) NOT NULL,
   status varchar(20) NOT NULL,
   PRIMARY KEY (id)
